refactor : 모달타입-상수분리
