+++ src/features/server/ui/ChannelItem.tsx
@@ -2,9 +2,10 @@ import { NavLink } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import { useMutation } from '@tanstack/react-query';
 import type { Channel } from '../model/types';
-import { MODAL, useModalStore } from '@/shared/model/modalStore';
+import { useModalStore } from '@/shared/model/modalStore';
 import { queryClient } from '@/shared/api/queryClient';
 import { QUERY_KEYS } from '@/shared/api/queryKeys';
+import { MODAL } from '@/shared/constants/modal';
 import { ROUTES } from '@/shared/constants/routes';
 import { AddUserIcon } from '@/shared/icons/AddUserIcon';
 import { ChannelUnlockedIcon } from '@/shared/icons/ChannelUnlockedIcon';
