+++ src/features/server/ui/EditServerModal.tsx
@@ -6,6 +6,7 @@ import { useModalStore } from '@/shared/model/modalStore';
 import useGetParams from '@/entities/hooks/getParams';
 import { queryClient } from '@/shared/api/queryClient';
 import { QUERY_KEYS } from '@/shared/api/queryKeys';
+import { MODAL } from '@/shared/constants/modal';
 import { ModalContainer } from '@/shared/ui/layout/ModalContainer';
 import { serverQueries } from '../api/queries';
 import { UploadImageInput } from './UploadImageInput';
@@ -68,7 +69,7 @@ export const EditServerModal = () => {
 
   return (
     <ModalContainer
-      isOpen={type === 'EDIT_SERVER'}
+      isOpen={type === MODAL.EDIT_SERVER}
       onClose={onCloseModal}
       subTitle="서버 개요"
       description=""
