+++ src/features/server/ui/JoinToServerModal.tsx
@@ -1,9 +1,10 @@
 import { toast } from 'react-toastify';
 import { useState } from 'react';
 import { useMutation } from '@tanstack/react-query';
-import { MODAL, useModalStore } from '@/shared/model/modalStore';
+import { useModalStore } from '@/shared/model/modalStore';
 import { queryClient } from '@/shared/api/queryClient';
 import { QUERY_KEYS } from '@/shared/api/queryKeys';
+import { MODAL } from '@/shared/constants/modal';
 import { ModalContainer } from '@/shared/ui/layout/ModalContainer';
 import { serverQueries } from '../api/queries';
