+++ src/features/server/ui/ServerDropDownMenu.tsx
@@ -1,5 +1,5 @@
-import { MODAL, useModalStore } from '@/shared/model/modalStore';
-import { ModalType } from '@/shared/constants/modalType';
+import { useModalStore } from '@/shared/model/modalStore';
+import { MODAL } from '@/shared/constants/modal';
 import { AddUserIcon } from '@/shared/icons/AddUserIcon';
 import { ChannelUnlockedIcon } from '@/shared/icons/ChannelUnlockedIcon';
 import { SettingIcon } from '@/shared/icons/SettingIcon';
@@ -22,15 +22,15 @@ export const ServerDropDownMenu = () => {
       <DropdownMenuLabel>서버 관리</DropdownMenuLabel>
       <DropdownMenuSeparator className="bg-gray opacity-50" />
       <DropdownMenuGroup>
-        <DropdownMenuItem onClick={() => handleModal(ModalType.INVIDE_MEMBER)}>
+        <DropdownMenuItem onClick={() => handleModal(MODAL.INVIDE_MEMBER)}>
           <span className="mr-auto">친구 초대하기</span>
           <AddUserIcon size={15} />
         </DropdownMenuItem>
-        <DropdownMenuItem onClick={() => handleModal(ModalType.EDIT_SERVER)}>
+        <DropdownMenuItem onClick={() => handleModal(MODAL.EDIT_SERVER)}>
           <span className="mr-auto">서버 설정</span>
           <SettingIcon size={15} />
         </DropdownMenuItem>
-        <DropdownMenuItem onClick={() => handleModal(ModalType.CREATE_CHANNEL)}>
+        <DropdownMenuItem onClick={() => handleModal(MODAL.CREATE_CHANNEL)}>
           <span className="mr-auto">채널 만들기</span>
           <ChannelUnlockedIcon size={15} />
         </DropdownMenuItem>
@@ -38,7 +38,7 @@ export const ServerDropDownMenu = () => {
       <DropdownMenuSeparator className="bg-gray opacity-50" />
       <DropdownMenuItem
         className="text-red focus:bg-red"
-        onClick={() => handleModal(ModalType.DELETE_SERVER)}
+        onClick={() => handleModal(MODAL.DELETE_SERVER)}
       >
         <span className="">서버 삭제하기</span>
       </DropdownMenuItem>
