+++ src/pages/channels/index.tsx
@@ -1,7 +1,8 @@
-import { MODAL, useModalStore } from '@/shared/model/modalStore';
+import { useModalStore } from '@/shared/model/modalStore';
 import { ChannelList } from '@/widgets/channel/ui/ChannelList';
 import { EditServerModal } from '@/features/server/ui/EditServerModal';
 import { InvitedMemberModal } from '@/features/server/ui/InvitedMemberModal';
+import { MODAL } from '@/shared/constants/modal';
 import { SEO_CONFIG } from '@/shared/constants/seo';
 import { Seo } from '@/shared/ui/Seo';
