+++ src/shared/constants/modal.ts
@@ -1,4 +1,4 @@
-export const ModalType = {
+export const MODAL = {
   CREATE_SERVER: 'CREATE_SERVER',
   EDIT_SERVER: 'EDIT_SERVER',
   DELETE_SERVER: 'DELETE_SERVER',
@@ -7,5 +7,7 @@ export const ModalType = {
   INVIDE_MEMBER: 'INVIDE_MEMBER',
   USER_PROFILE: 'USER_PROFILE',
   USER_SIGNOFF: 'USER_SIGNOFF',
+  JOIN_SERVER: 'JOIN_SERVER',
+  CALL_NOTIFICATION: 'CALL_NOTIFICATION',
   NONE: 'NONE',
 } as const;
