+++ src/shared/model/modalStore.ts
@@ -1,18 +1,5 @@
 import { create } from 'zustand';
-
-export const MODAL = {
-  CREATE_SERVER: 'CREATE_SERVER',
-  EDIT_SERVER: 'EDIT_SERVER',
-  DELETE_SERVER: 'DELETE_SERVER',
-  CREATE_CHANNEL: 'CREATE_CHANNEL',
-  DELETE_CHANNEL: 'DELETE_CHANNEL',
-  INVIDE_MEMBER: 'INVIDE_MEMBER',
-  USER_PROFILE: 'USER_PROFILE',
-  USER_SIGNOFF: 'USER_SIGNOFF',
-  JOIN_SERVER: 'JOIN_SERVER',
-  CALL_NOTIFICATION: 'CALL_NOTIFICATION',
-  NONE: 'NONE',
-} as const;
+import { MODAL } from '../constants/modal';
 
 export interface ModalStore {
   type: keyof typeof MODAL;
